backend/requirements.txt
fastapi
uvicorn
pydantic>=2.5,<3
pydantic-ai
python-dotenv
PyPDF2
python-multipart
httpx
